+++ yalafi/packages/__init__.py
@@ -7,6 +7,7 @@ load_table = {
         'amsthm',
         'babel',
         'biblatex',
+        'circuitikz',
         # 'cleveref',
         # cleveref is not loaded by default because it throws
         # a warning if loaded without the poorman option.