+++ yalafi/packages/amsmath.py
@@ -3,7 +3,7 @@
 #   YaLafi module for LaTeX package amsmath
 #
 
-from yalafi.defs import Macro, InitModule, EquEnv
+from yalafi.defs import Macro, InitModule, EquEnv, SpecialToken
 
 require_packages = []
 
@@ -15,7 +15,7 @@ require_packages = []
 #   - The spacing macros and \notag are imporant for correct parsing
 #     of maths material.
 #
-def init_module(parser, options):
+def init_module(parser, options, position):
     parms = parser.parms
 
     parms.math_text_macros.append('\\text')
@@ -37,6 +37,7 @@ def init_module(parser, options):
     macros_python = [
 
         Macro(parms, '\\DeclareMathOperator', args='*AA'),
+        Macro(parms, '\\substack', args='A', repl=h_substack),
 
     ]
 
@@ -57,6 +58,19 @@ def init_module(parser, options):
 
     ]
 
+    inject_tokens = []      # this list of tokens is used as replacement
+                            # for '\documentclass...' or '\usepackage...'
+
     return InitModule(macros_latex=macros_latex, macros_python=macros_python,
-                        environments=environments)
+                    environments=environments, inject_tokens=inject_tokens)
+
+
+#   macro \substack: replace \\ tokens with space
+#
+def h_substack(parser, buf, mac, args, delim, pos):
+    def f(tok, lev):
+        if tok.txt == r'\\' and lev == 0:
+            return SpecialToken(tok.pos, r'\;')
+        return tok
+    return [f(tok, lev) for tok, lev in parser.iter_token_levels(args[0])]