+++ yalafi/packages/babel.py
@@ -65,7 +65,7 @@ from yalafi.defs import InitModule, Macro, LanguageToken, Environ, MacroToken
 
 require_packages = []
 
-def init_module(parser, options):
+def init_module(parser, options, position):
     parms = parser.parms
 
     macros_latex = ''