+++ yalafi/packages/mathtools.py
@@ -8,7 +8,7 @@ from yalafi.defs import InitModule
 require_packages = ["amsmath"]
 
 
-def init_module(parser, options):
+def init_module(parser, options, position):
     parms = parser.parms
 
     macros_latex = r"""