+++ yalafi/packages/xspace.py
@@ -8,7 +8,7 @@ from yalafi.defs import Macro, InitModule
 
 require_packages = []
 
-def init_module(parser, options):
+def init_module(parser, options, position):
     parms = parser.parms
 
     macros_latex = ""